requirements.txt
openai
python-dotenv
schedule
requests
transformers
torch
sentencepiece
flask 
accelerate